write a bit of documentation for the plugin
